refactor: 관리자 예약 기능을 별도 인터페이스로 분리
