+++ src/main/java/roomescape/feature/reservation/service/AdminReservationService.java
@@ -1,0 +1,11 @@
+package roomescape.feature.reservation.service;
+
+import java.util.List;
+import roomescape.feature.reservation.dto.response.ReservationResponseDto;
+
+public interface AdminReservationService {
+
+    List<ReservationResponseDto> getReservations();
+
+    void deleteReservationById(Long id);
+}
